feat(group_theory/perm/concrete_cycle): cyclic perm notation
## Main definitions

In the following, `{α : Type*} [fintype α] [decidable_eq α]`.

* `cycle.form_perm`: the cyclic permutation created by looping over a `cycle α`
* `equiv.perm.to_list`: the `list` representation of the cycle generated by
  `f : perm α` starting from `x : α`
* `equiv.perm.to_cycle`: the `cycle` representation of the cycle generated by
  `f : perm α` where `hf : f.is_cycle`
* `equiv.perm.iso_cycle`: the equivalence between cyclic permutations `f : perm α`
  and the terms of `cycle α` that correspond to them
* `equiv.perm.iso_cycle'`: the same equivalence as `equiv.perm.iso_cycle`
  but with faster evaluation via `#eval`
* The notation `c[1, 2, 3]` to emulate notation of cyclic permutations `(1 2 3)`
* A `has_repr` instance for any `perm α`, by representing the `finset` of
  `cycle α` that correspond to the cycle factors.

## Main results

* `equiv.perm.is_cycle.exists_unique_cycle`: there is only one nontrivial `cycle α`
  corresponding to each cyclic `f : perm α`

## Implementation details

The forward direction of `equiv.perm.iso_cycle` uses `fintype.choose` of the uniqueness
result, relying on the `fintype` instance of a `cycle.nodup` subtype.
For some reason, this works faster than the `equiv.perm.to_cycle`, which relies
on recursion over `finset.univ`.

Running `#eval` on even a simple noncyclic permutation `c[(1 : fin 7), 2, 3] * c[0, 5]`
to show it takes a long time. TODO: is this because computing the cycle factors is slow?